requirements.txt
khl.py==0.3.17
flask==2.0.1
werkzeug==2.0.3
requests==2.26.0
python-dotenv==0.19.0
simple-kook-voice==0.2.6